refactor: get rid of the pesky lib and move the code to writer.py
